backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.37
asyncpg==0.30.0
alembic==1.14.1

# Data Validation
pydantic[email]==2.10.6
pydantic-settings==2.8.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Config
python-dotenv==1.0.1

# Task Queue
celery==5.4.0
redis==5.2.1

# LLM / Agents
langchain==0.3.18
langchain-groq==0.2.4

# Push Notifications
pywebpush==2.0.0
